Socket name can be forced by exporting IBSIM_SOCKNAME before starting ibsim and/or preloading umad2sim so multiple simulator can run on the same system at the same time
As we do a lot of routing tests with ibsim we had the need to be able to launch multiple simulator on the same system.
With this patch, ibsim (and umad2sim) will try to read the socket basename using a getenv("IBSIM_SOCKNAME") which makes it possible.
If IBSIM_SOCKNAME is not set, SIM_BASENAME is still used.

Signed-off-by: Nicolas Morey-Chaisemartin <nicolas.morey-chaisemartin@ext.bull.net>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>