ibsim/sim_net.c: Assume QDR speed when port speed is 0
If there are devices with active extended speed, that responds to
PortInfo with LinkSpeedActive set to 0 (instead of QDR)
ibnetdiscover -f writes links with LinkSpeedActive set to 0 (s=0),
and as a result ibsim is unable to parse the file.

With this change, ibsim assumes LinkSpeedActive as QDR when
LinkSpeedActive is 0.

Signed-off-by: Daniel Klein <dklein@nvidia.com>